chore(conftest): agregar fixtures a nivel general: usuario_cliente y usuario_profesional
